export banana split passphrase manu added and bottom sheet clarifications
